requirements.txt
sounddevice
numpy
pillow
tcl-tk
librosa
matplotlib
